Move stuff around and rename things.
